TOOLS/PERF: workaround for cuda_ipc UB
The cuda_ipc UCT relies on a CUDA UB to cache an opened IPC memhandle so
it can avoid reopening it.
However, perftest's main loop free its src/dst buffers after each loop
when the buffer's IPC handle is still opened and cached in UCT.
This leads to an UB in CUDA---calling free before closing the mapped
handle.